refactor: RestaurantDataUpdater 인덱스 추가 및 기존 경로 수정
